refactor graphiql explorer into multiple files (#3375)
