+++ packages/graphiql-explorer/src/index.ts
@@ -1,0 +1,2 @@
+export type { Props as GraphiQLExplorerProps } from './explorer';
+export { Explorer } from './explorer';
