+++ packages/graphiql-explorer/tsconfig.json
@@ -22,5 +22,5 @@
     "baseUrl": ".",
     "rootDir": "src"
   },
-  "include": ["src"]
+  "include": ["src/**/*"]
 }
